\name{afExp}
\alias{afExp}
\docType{data}
\title{Gene expression profile}
\description{
A matrix of data specifying for the gene set expression profile, each row presents a expression profile of a specific gene, each column represents a sample.
}
\usage{data(afExp)}
\format{
These are both gene expression profiles, each row represents a gene, each column represents a sample

}
\source{
They are based on the data obtained as followed, by deleting the first row and first column
http://www.broadinstitute.org/gsea/msigdb/index.jsp

}
\examples{
data(afExp)
}
\keyword{datasets}
