\name{b1}
\alias{b1}
\docType{data}
\title{Gene set function expression profile}
\description{
This is a data file containing three items, gene function label, gene name, gene expression profile, a function label may contain many genes , each gene contains many expression profiles, each row has a gene function label, a gene name and the gene's expression profile, each column represents a sample.    

}

\source{
These data are obtained from original data by some operations.
}
\examples{
filename=system.file("data", "b1.txt", package="DBGSA")
}
\keyword{datasets}

