\name{DBGSA-package}
\alias{DBGSA-package}
\alias{DBGSA}
\docType{package}
\title{methods of distance-based gene set functional enrichment analysis}
\description{DBGSA package provides a new method to evaluate the possiblities of the enrichment between different phenotypes, which is based on the distances between them. A fdr method is used to adjust the p value.}
\details{
\tabular{ll}{
Package: \tab DBGSA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Depends: \tab R(>=2.12.2),methods,fdrtool\cr
Suggest: \tab ALL\cr
Date: \tab 2011-03-23\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
BiocViews: \tab bioinformatics\cr
}

}

\author{ 
Li Jin, Huang Meilin
Maintainer:Li Jin<lijin@hrbmu.edu.cn>
}
\references{
Lu Tian, Steven A. Greenberg, Sek Won Kong, Josiah Altschuler,Discovering statistically significant pathways in expression profiling studies, PNAS 2005 102;13544-13549
}

\keyword{ package }
