\name{distance}
\alias{distance}
\docType{data}
\title{distances between case and control group}
\description{
This is a data file that has three columns: gene function label, the number of gene the former gene function label contains, distances between case and control group in this gene function lable, each row of this matrix represents the information of a gene function label. 
 
}
\usage{data(distance)}

\source{
these data are obtained by the distable function, and then delete the first row
}

\examples{
data(distance)

}
\keyword{datasets}
