\name{rand}
\alias{rand}
\docType{data}
\title{distances of gene expression obtained by gene resampling}
\description{
An array of data specifying for distances of the gene set expression profile by gene resampling, each row presents a expression profile of a specific gene label, each column represents a sample.
}
\usage{data(rand)}
\format{
These are both distances of gene expression profiles, each row represents a gene label, each column represents a sample

}
\source{
They are derived unsing \code{randdis}

}
\examples{
data(rand)
}
\keyword{datasets}
