\name{DBGSA-package}
\alias{DBGSA-package}
\alias{DBGSA}
\docType{package}
\title{methods of distance-based gene set functional enrichment analysis}
\description{DBGSA package provides a new method to evaluate the possiblities of the enrichment between different phenotypes. DBGSA is a novel distance-based gene set  enrichment analysis method. We consider that, the distance between 2 groups with  different phenotype by focusing on the gene expression should be larger, if a  certain gene functional set was significantly associated with a particular  phenotype.}
\details{
\tabular{ll}{
Package: \tab DBGSA\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Depends: \tab R(>=2.12.2),methods,fdrtool\cr
Suggest: \tab ALL\cr
Date: \tab 2011-05-17\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
BiocViews: \tab bioinformatics\cr
}

}

\author{ 
Li Jin, Huang Meilin
Maintainer:Li Jin<lijin@hrbmu.edu.cn>
}
\references{
Lu Tian, Steven A. Greenberg, Sek Won Kong, Josiah Altschuler,Discovering statistically significant pathways in expression profiling studies, PNAS 2005 102;13544-13549
}

\keyword{ package }
